refactor: Extract shared types to _types.py
- Move AtomData and SasaResultWithAtoms to shared _types.py module
- Update gemmi.py to import from _types.py and re-export for backward compatibility
- Update biopython.py to import from _types.py instead of gemmi
- Add docstring about hydrogen detection limitations
